+++ src/Command/CacheClearCommand.php
@@ -11,7 +11,6 @@
 
 namespace Gnugat\MicroFrameworkBundle\Command;
 
-use Gnugat\MicroFrameworkBundle\Console\ExitCode;
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
@@ -40,10 +39,10 @@ class CacheClearCommand extends Command
                 php bin/console cache:clear
                 php bin/console cache:warmup
             HELP
-            );
+        );
     }
 
-    protected function execute(InputInterface $input, OutputInterface $output)
+    protected function execute(InputInterface $input, OutputInterface $output): int
     {
         $io = new SymfonyStyle($input, $output);
         $io->comment("Clearing the cache (<info>{$this->cacheDir}</info>)");
@@ -52,6 +51,6 @@ class CacheClearCommand extends Command
 
         $io->success('Cache cleared');
 
-        return ExitCode::SUCCESS;
+        return Command::SUCCESS;
     }
 }