+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # CHANGELOG
 
+## 0.9.0: Symfony 7 support
+
+* Upgraded support to Symfony 6
+* Upgraded support to PHP 8.2 (accordingly with Symfony 7)
+* Upgraded tests to PHPUnit 11
+
+- Removed ExitCode (use Symfony\Component\Console\Command constants instead)
+- Removed support for `ContainerAwareCommand`, as it's been removed from Symfony 7
+
 ## 0.8.0: Symfony 6 support
 
 * Upgraded support to Symfony 6